remove unnecessary cast.

+++ Sources/Libraries/EigenWrapper/Sample.h
@@ -10,6 +10,7 @@
 
 @interface Sample : NSObject
 
-+ (void)multiple:(int *)result :(int)aRow :(int)aCol :(int)bCol :(const int[])aGrid :(const int[])bGrid;
++ (void)multiple:(NSInteger *)result :(NSInteger)aRow :(NSInteger)aCol :(NSInteger)bCol :(const NSInteger[])aGrid :(const NSInteger[])bGrid;
 
 @end
+
